+++ projects/ngx-pk-ui/src/lib/pk-timepicker/pk-timepicker.model.ts
@@ -1,0 +1,3 @@
+export type PkTimeFormat = 'hms' | 'hm' | 'h';
+export type PkTimeType = '24H' | '12H';
+export type PkTimeInputType = 'spinner' | 'number' | 'dropdown';